requirements.txt
python-xlib==0.31
six==1.16.0
